refactor: move ui-camera module to style components modules.
