+++ features/infos/infos-ui/build.gradle.kts
@@ -28,7 +28,7 @@ kotlin {
         }
         val androidMain by getting {
             dependencies {
-                implementation(projects.uiCamera)
+                implementation(projects.style.components.camera)
             }
         }
     }
