+++ features/infos/infos-ui/src/androidMain/kotlin/com/paligot/confily/infos/ui/tickets/TicketCameraPreview.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import com.paligot.confily.ui.camera.CameraPreview
+import com.paligot.confily.style.components.camera.CameraPreview
 
 @Composable
 fun TicketCameraPreview(
