+++ features/networking/networking-ui/build.gradle.kts
@@ -21,7 +21,7 @@ kotlin {
         }
         val androidMain by getting {
             dependencies {
-                implementation(projects.uiCamera)
+                implementation(projects.style.components.camera)
             }
         }
     }
