+++ settings.gradle.kts
@@ -91,6 +91,7 @@ include(":shared:models")
 include(":shared:resources")
 include(":backend")
 include(":style:components:adaptive")
+include(":style:components:camera")
 include(":style:components:markdown")
 include(":style:components:permissions")
 include(":style:components:placeholder")
@@ -100,7 +101,6 @@ include(":style:partners")
 include(":style:schedules")
 include(":style:speakers")
 include(":style:theme")
-include(":ui-camera")
 include(":webApps:speakerApp")
 include(":wear:wear-app")
 include(":wear:wear-features:events:events-di")
